Progress towards api routes
