+++ express-back-end/package.json
@@ -9,7 +9,7 @@
   "scripts": {
     "go": "nodemon dist/server.js",
     "db:reset": "node dist/bin/resetdb.js",
-    "db:reset-seeds": "node dist/bin/generate-entry-seeds.js " 
+    "db:reset-seeds": "node dist/bin/generate-entry-seeds.js "
   },
   "dependencies": {
     "body-parser": "^1.18.3",
